Do all audio processing in 32 bit floating point

+++ include/band.h
@@ -17,7 +17,7 @@ int band_set_channel_synth(int channel, const SynthType *type);
 
 const char **band_get_channel_controls(int channel, int *numControls);
 float *band_get_channel_control(int channel, const char *control);
-void band_run(int16_t *buffer, int length);
+void band_run(float *buffer, int length);
 
 bool band_send_note(uint32_t time, int channel, bool state, int num, float velocity);
 bool band_send_pitch(uint32_t time, int channel, float offset);
